:sparkles: Update: bug fixes + en language

+++ HydrateMe/app/src/main/java/com/example/hydrateme/waterfall/Helpers.kt
@@ -4,10 +4,10 @@ import android.content.Context
 import android.util.TypedValue
 import java.io.Serializable
 
-fun getFormula(sex: Byte): (Float, Float) -> Double {
+fun getFormula(sex: Byte): (Float, Float) -> Float {
     when (sex) {
-        0.toByte() -> return { weight: Float, actTime: Float -> weight * 0.03 + actTime * 0.5 }
-        1.toByte() -> return { weight: Float, actTime: Float -> weight * 0.025 + actTime * 0.4 }
+        0.toByte() -> return { weight: Float, actTime: Float -> weight * 0.03F + actTime * 0.5F }
+        1.toByte() -> return { weight: Float, actTime: Float -> weight * 0.025F + actTime * 0.4F }
         else -> throw InvalidArgumentException("Sex argument can be 0 or 1. $sex found")
     }
 }
